+++ plugin/ui/textboxlitewidget.lua
@@ -37,16 +37,13 @@ function TextBoxLiteWidget:init()
 end
 
 function TextBoxLiteWidget:setText(text)
-    if not self._center_abs then
-        -- no cached center => we haven't drawn anything yet
-        TextWidget.setText(self, text)
-        self:free()
-        return
-    end
+    if self.text == text then return end
     local center_abs = self._center_abs
     local prev_text_region = self._text_region_abs
-    TextWidget.setText(self, text)
-    self:free(false)
+    self:clearCache(false)
+    self.text = text
+    -- no cached center => we haven't drawn anything yet
+    if not center_abs then return end
     local next_text_region = self:centerAlignedTextBounds(center_abs, TextWidget.getSize(self))
     local clear_region = Geom.boundingBox({ prev_text_region, next_text_region })
     UIManager:setDirty(self.ui_root, self.refreshtype, clear_region)
@@ -62,13 +59,23 @@ function TextBoxLiteWidget:getSize()
     return self._dimen
 end
 
-function TextBoxLiteWidget:free(drop_position)
+function TextBoxLiteWidget:setWidth(width)
+    self.width = width
+    self:clearCache(true)
+    self.max_width = self.width - 2 * self.box_padding
+end
+
+function TextBoxLiteWidget:clearCache(drop_position)
     if drop_position then
         self._center_abs = nil
         self._text_region_abs = nil
     end
     self._dimen = nil
     TextWidget.free(self)
+end
+
+function TextBoxLiteWidget:free()
+    self:clearCache(true)
 end
 
 function TextBoxLiteWidget:centerAlignedTextBounds(center_abs, text_size)