+++ CHANGELOG.md
@@ -5,6 +5,24 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 [//]: # (## [Unreleased])
 
+## [2025.12-1] - 2025-12-22
+
+This small update adds some status information (time, battery, frontlight) to
+the bottom of the panel. It also includes a render fix after device deep sleep.
+
+Merry Christmas 🎄 and a happy new year! 🚀
+
+### Added
+
+- [#8](https://github.com/oleasteo/koreader-screenlockpin/issues/8)
+  Device status text with time, battery, and frontlight. 
+
+### Fixed
+
+- [#38](https://github.com/oleasteo/koreader-screenlockpin/issues/38)
+  Flash initial lock screen panel draw to avoid a transparent background by
+  accident (only black is drawn, not white). 
+
 ## [2025.12] - 2025-12-17
 
 If everything works out, this should be the last update that you have to do by